Interactive: make module to check input
