+++ src/modules/UI.js
@@ -1,16 +1,17 @@
 /* eslint-disable eqeqeq */
 
-import Todo from './todo.js';
-import sortArray from '../../node_modules/sort-array/dist/index.mjs';
+import Todo from "./todo.js";
+import sortArray from "../../node_modules/sort-array/dist/index.mjs";
+import preValidate from "./completed.js";
 
-const dataInput = document.querySelector('.data_input');
-const todoList = document.querySelector('.todo_list');
+const dataInput = document.querySelector(".data_input");
+const todoList = document.querySelector(".todo_list");
 
 let todoItems = [];
 
 export default class UI {
   static getItems() {
-    todoItems = JSON.parse(localStorage.getItem('todoItems'));
+    todoItems = JSON.parse(localStorage.getItem("todoItems"));
     if (todoItems) return todoItems;
     return [];
   }
@@ -22,13 +23,13 @@ export default class UI {
     filterTodoItems.forEach((element) => {
       if (element.index > ID) element.index -= 1;
     });
-    localStorage.setItem('todoItems', JSON.stringify(filterTodoItems));
+    localStorage.setItem("todoItems", JSON.stringify(filterTodoItems));
     window.location.reload();
   }
 
   static isChecked(check) {
-    if (check === true) return 'checked';
-    return 'notChecked';
+    if (check === true) return "checked";
+    return "notChecked";
   }
 
   static check() {
@@ -37,7 +38,7 @@ export default class UI {
     todoItems.forEach((elem) => {
       if (elem.completed == true) {
         document.querySelector(
-          `[data-id="${elem.index}"] > .todo_check`,
+          `[data-id="${elem.index}"] > .todo_check`
         ).checked = true;
       }
     });
@@ -46,11 +47,11 @@ export default class UI {
   static displayItems() {
     todoItems = this.getItems();
     const sortedArray = sortArray(todoItems, {
-      by: 'openIssues',
+      by: "openIssues",
     });
     todoItems = sortedArray;
     if (todoItems != null) {
-      todoList.innerHTML = '';
+      todoList.innerHTML = "";
       todoItems.forEach((element) => {
         todoList.innerHTML += ` <li data-id="${element.index}" data-valid="${element.completed}"><input type="checkbox"  class="todo_check" />
                 <input type="text" value="${element.description}"class="todo_input" />
@@ -67,56 +68,39 @@ export default class UI {
       const todoItem = new Todo(dataInput.value, false, todoItems.length + 1);
 
       todoItems.push(todoItem);
-      localStorage.setItem('todoItems', JSON.stringify(todoItems));
+      localStorage.setItem("todoItems", JSON.stringify(todoItems));
     }
   }
 
   static deleteItem(e) {
-    const dataId = e.target.parentElement.getAttribute('data-id');
+    const dataId = e.target.parentElement.getAttribute("data-id");
     this.filterByID(dataId);
   }
 
   static changeItem(e, val) {
-    const dataId = e.target.parentElement.getAttribute('data-id');
+    const dataId = e.target.parentElement.getAttribute("data-id");
 
     todoItems = this.getItems();
 
     todoItems.forEach((elm) => {
       if (+elm.index === +dataId) elm.description = val;
     });
-    localStorage.setItem('todoItems', JSON.stringify(todoItems));
+    localStorage.setItem("todoItems", JSON.stringify(todoItems));
     window.location.reload();
   }
 
   static clearCompleted(e) {
-    const dataID = e.closest('li').getAttribute('data-id');
+    const dataID = e.closest("li").getAttribute("data-id");
     this.filterByID(dataID);
   }
 
-  static preValidate(valid, e) {
-    const dataId = e.target.parentElement.getAttribute('data-id');
-
-    e.target.parentElement.setAttribute('data-valid', valid);
+  static validate(e) {
+    const isValid = e.target.parentElement.getAttribute("data-valid");
 
     todoItems = this.getItems();
-    todoItems.find((elem) => {
-      let i;
-      if (elem.index == dataId) {
-        elem.completed = valid;
-        i = 1;
-      } else i = 0;
-      return i;
-    });
-
-    localStorage.setItem('todoItems', JSON.stringify(todoItems));
-  }
-
-  static validate(e) {
-    const isValid = e.target.parentElement.getAttribute('data-valid');
-
-    if (isValid == 'false') {
-      this.preValidate(true, e);
-    } else this.preValidate(false, e);
+    if (isValid == "false") {
+      preValidate(true, e, todoItems);
+    } else preValidate(false, e, todoItems);
   }
 
   // end
